fc_rpc_core

Trait DebugApiServer

pub trait DebugApiServer:
    Sized
    + Send
    + Sync
    + 'static {
    // Required methods
    fn raw_header<'life0, 'async_trait>(
        &'life0 self,
        number: BlockNumberOrHash,
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Bytes>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn raw_block<'life0, 'async_trait>(
        &'life0 self,
        number: BlockNumberOrHash,
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Bytes>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn raw_transaction<'life0, 'async_trait>(
        &'life0 self,
        hash: H256,
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Bytes>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn raw_receipts<'life0, 'async_trait>(
        &'life0 self,
        number: BlockNumberOrHash,
    ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<Bytes>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn bad_blocks(&self, number: BlockNumberOrHash) -> RpcResult<Vec<()>>;

    // Provided method
    fn into_rpc(self) -> RpcModule<Self> { ... }
}

Server trait implementation for the DebugApi RPC API.

Required Methods

fn raw_header<'life0, 'async_trait>( &'life0 self, number: BlockNumberOrHash, ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Bytes>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns an RLP-encoded header with the given number or hash.

fn raw_block<'life0, 'async_trait>( &'life0 self, number: BlockNumberOrHash, ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Bytes>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns an RLP-encoded block with the given number or hash.

fn raw_transaction<'life0, 'async_trait>( &'life0 self, hash: H256, ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Bytes>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns a EIP-2718 binary-encoded transaction with the given hash.

fn raw_receipts<'life0, 'async_trait>( &'life0 self, number: BlockNumberOrHash, ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<Bytes>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns an array of EIP-2718 binary-encoded receipts with the given number of hash.

fn bad_blocks(&self, number: BlockNumberOrHash) -> RpcResult<Vec<()>>

Returns an array of recent bad blocks that the client has seen on the network.

Provided Methods

fn into_rpc(self) -> RpcModule<Self>

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.