fc_rpc_core::types

Struct Transaction

pub struct Transaction {

    pub transaction_type: U256,
    pub hash: H256,
    pub nonce: U256,
    pub block_hash: Option<H256>,
    pub block_number: Option<U256>,
    pub transaction_index: Option<U256>,
    pub from: H160,
    pub to: Option<H160>,
    pub value: U256,
    pub gas: U256,
    pub gas_price: Option<U256>,
    pub max_fee_per_gas: Option<U256>,
    pub max_priority_fee_per_gas: Option<U256>,
    pub input: Bytes,
    pub creates: Option<H160>,
    pub chain_id: Option<U64>,
    pub access_list: Option<Vec<AccessListItem>>,
    pub y_parity: Option<U256>,
    pub v: Option<U256>,
    pub r: U256,
    pub s: U256,

}

Transaction

Fields
transaction_type: U256

EIP-2718 transaction type

hash: H256

Hash

nonce: U256

Nonce

block_hash: Option<H256>

Block hash

block_number: Option<U256>

Block number

transaction_index: Option<U256>

Transaction Index

from: H160

Sender

to: Option<H160>

Recipient

value: U256

Transferred value

gas: U256

Gas

gas_price: Option<U256>

Gas Price

max_fee_per_gas: Option<U256>

Max BaseFeePerGas the user is willing to pay.

max_priority_fee_per_gas: Option<U256>

The minerโs tip.

input: Bytes

Data

creates: Option<H160>

Creates contract

chain_id: Option<U64>

The network id of the transaction, if any.

access_list: Option<Vec<AccessListItem>>

Pre-pay to warm storage access.

y_parity: Option<U256>

The parity (0 for even, 1 for odd) of the y-value of the secp256k1 signature.

v: Option<U256>

The standardised V field of the signature.

For backwards compatibility, v is optionally provided as an alternative to yParity. This field is DEPRECATED and all use of it should migrate to yParity.

r: U256

The R field of the signature.

s: U256

The S field of the signature.