Trait fc_rpc_core::EthApiServer

pub trait EthApiServer: Sized + Send + Sync + 'static {

    // Required methods
    fn protocol_version(&self) -> RpcResult<u64>;
    fn syncing<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = RpcResult<SyncStatus>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn author(&self) -> RpcResult<H160>;
    fn accounts(&self) -> RpcResult<Vec<H160>>;
    fn block_number(&self) -> RpcResult<U256>;
    fn chain_id(&self) -> RpcResult<Option<U64>>;
    fn block_by_hash<'life0, 'async_trait>(
        &'life0 self,
        hash: H256,
        full: bool
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<RichBlock>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn block_by_number<'life0, 'async_trait>(
        &'life0 self,
        number: BlockNumber,
        full: bool
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<RichBlock>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn block_transaction_count_by_hash<'life0, 'async_trait>(
        &'life0 self,
        hash: H256
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<U256>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn block_transaction_count_by_number<'life0, 'async_trait>(
        &'life0 self,
        number: BlockNumber
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<U256>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn block_transaction_receipts<'life0, 'async_trait>(
        &'life0 self,
        number: BlockNumber
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Vec<Receipt>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn block_uncles_count_by_hash(&self, hash: H256) -> RpcResult<U256>;
    fn block_uncles_count_by_number(
        &self,
        number: BlockNumber
    ) -> RpcResult<U256>;
    fn uncle_by_block_hash_and_index(
        &self,
        hash: H256,
        index: Index
    ) -> RpcResult<Option<RichBlock>>;
    fn uncle_by_block_number_and_index(
        &self,
        number: BlockNumber,
        index: Index
    ) -> RpcResult<Option<RichBlock>>;
    fn transaction_by_hash<'life0, 'async_trait>(
        &'life0 self,
        hash: H256
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Transaction>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn transaction_by_block_hash_and_index<'life0, 'async_trait>(
        &'life0 self,
        hash: H256,
        index: Index
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Transaction>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn transaction_by_block_number_and_index<'life0, 'async_trait>(
        &'life0 self,
        number: BlockNumber,
        index: Index
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Transaction>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn transaction_receipt<'life0, 'async_trait>(
        &'life0 self,
        hash: H256
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Receipt>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn balance<'life0, 'async_trait>(
        &'life0 self,
        address: H160,
        number: Option<BlockNumber>
    ) -> Pin<Box<dyn Future<Output = RpcResult<U256>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn storage_at<'life0, 'async_trait>(
        &'life0 self,
        address: H160,
        index: U256,
        number: Option<BlockNumber>
    ) -> Pin<Box<dyn Future<Output = RpcResult<H256>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn transaction_count<'life0, 'async_trait>(
        &'life0 self,
        address: H160,
        number: Option<BlockNumber>
    ) -> Pin<Box<dyn Future<Output = RpcResult<U256>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn code_at<'life0, 'async_trait>(
        &'life0 self,
        address: H160,
        number: Option<BlockNumber>
    ) -> Pin<Box<dyn Future<Output = RpcResult<Bytes>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn call<'life0, 'async_trait>(
        &'life0 self,
        request: CallRequest,
        number: Option<BlockNumber>,
        state_overrides: Option<BTreeMap<H160, CallStateOverride>>
    ) -> Pin<Box<dyn Future<Output = RpcResult<Bytes>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn estimate_gas<'life0, 'async_trait>(
        &'life0 self,
        request: CallRequest,
        number: Option<BlockNumber>
    ) -> Pin<Box<dyn Future<Output = RpcResult<U256>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn gas_price(&self) -> RpcResult<U256>;
    fn fee_history<'life0, 'async_trait>(
        &'life0 self,
        block_count: U256,
        newest_block: BlockNumber,
        reward_percentiles: Option<Vec<f64>>
    ) -> Pin<Box<dyn Future<Output = RpcResult<FeeHistory>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn max_priority_fee_per_gas(&self) -> RpcResult<U256>;
    fn is_mining(&self) -> RpcResult<bool>;
    fn hashrate(&self) -> RpcResult<U256>;
    fn work(&self) -> RpcResult<Work>;
    fn submit_hashrate(&self, hashrate: U256, id: H256) -> RpcResult<bool>;
    fn submit_work(
        &self,
        nonce: H64,
        pow_hash: H256,
        mix_digest: H256
    ) -> RpcResult<bool>;
    fn send_transaction<'life0, 'async_trait>(
        &'life0 self,
        request: TransactionRequest
    ) -> Pin<Box<dyn Future<Output = RpcResult<H256>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn send_raw_transaction<'life0, 'async_trait>(
        &'life0 self,
        bytes: Bytes
    ) -> Pin<Box<dyn Future<Output = RpcResult<H256>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

    // Provided method
    fn into_rpc(self) -> RpcModule<Self> { ... }

}

Server trait implementation for the EthApi RPC API.

Required Methods

fn protocol_version(&self) -> RpcResult<u64>

Returns protocol version encoded as a string (quotes are necessary).

fn syncing<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = RpcResult<SyncStatus>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Returns an object with data about the sync status or false. (wtf?)

fn author(&self) -> RpcResult<H160>

Returns block author.

fn accounts(&self) -> RpcResult<Vec<H160>>

Returns accounts list.

fn block_number(&self) -> RpcResult<U256>

Returns highest block number.

fn chain_id(&self) -> RpcResult<Option<U64>>

Returns the chain ID used for transaction signing at the current best block. None is returned if not available.

fn block_by_hash<'life0, 'async_trait>( &'life0 self, hash: H256, full: bool ) -> Pin<Box<dyn Future<Output = RpcResult<Option<RichBlock>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Returns block with given hash.

fn block_by_number<'life0, 'async_trait>( &'life0 self, number: BlockNumber, full: bool ) -> Pin<Box<dyn Future<Output = RpcResult<Option<RichBlock>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Returns block with given number.

fn block_transaction_count_by_hash<'life0, 'async_trait>( &'life0 self, hash: H256 ) -> Pin<Box<dyn Future<Output = RpcResult<Option<U256>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Returns the number of transactions in a block with given hash.

fn block_transaction_count_by_number<'life0, 'async_trait>( &'life0 self, number: BlockNumber ) -> Pin<Box<dyn Future<Output = RpcResult<Option<U256>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Returns the number of transactions in a block with given block number.

fn block_transaction_receipts<'life0, 'async_trait>( &'life0 self, number: BlockNumber ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Vec<Receipt>>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Returns the receipts of a block by number or hash.

fn block_uncles_count_by_hash(&self, hash: H256) -> RpcResult<U256>

Returns the number of uncles in a block with given hash.

fn block_uncles_count_by_number(&self, number: BlockNumber) -> RpcResult<U256>

Returns the number of uncles in a block with given block number.

fn uncle_by_block_hash_and_index( &self, hash: H256, index: Index ) -> RpcResult<Option<RichBlock>>

Returns an uncles at given block and index.

fn uncle_by_block_number_and_index( &self, number: BlockNumber, index: Index ) -> RpcResult<Option<RichBlock>>

Returns an uncles at given block and index.

fn transaction_by_hash<'life0, 'async_trait>( &'life0 self, hash: H256 ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Transaction>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get transaction by its hash.

fn transaction_by_block_hash_and_index<'life0, 'async_trait>( &'life0 self, hash: H256, index: Index ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Transaction>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Returns transaction at given block hash and index.

fn transaction_by_block_number_and_index<'life0, 'async_trait>( &'life0 self, number: BlockNumber, index: Index ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Transaction>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Returns transaction by given block number and index.

fn transaction_receipt<'life0, 'async_trait>( &'life0 self, hash: H256 ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Receipt>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Returns transaction receipt by transaction hash.

fn balance<'life0, 'async_trait>( &'life0 self, address: H160, number: Option<BlockNumber> ) -> Pin<Box<dyn Future<Output = RpcResult<U256>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Returns balance of the given account.

fn storage_at<'life0, 'async_trait>( &'life0 self, address: H160, index: U256, number: Option<BlockNumber> ) -> Pin<Box<dyn Future<Output = RpcResult<H256>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Returns content of the storage at given address.

fn transaction_count<'life0, 'async_trait>( &'life0 self, address: H160, number: Option<BlockNumber> ) -> Pin<Box<dyn Future<Output = RpcResult<U256>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Returns the number of transactions sent from given address at given time (block number).

fn code_at<'life0, 'async_trait>( &'life0 self, address: H160, number: Option<BlockNumber> ) -> Pin<Box<dyn Future<Output = RpcResult<Bytes>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Returns the code at given address at given time (block number).

fn call<'life0, 'async_trait>( &'life0 self, request: CallRequest, number: Option<BlockNumber>, state_overrides: Option<BTreeMap<H160, CallStateOverride>> ) -> Pin<Box<dyn Future<Output = RpcResult<Bytes>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Call contract, returning the output data.

fn estimate_gas<'life0, 'async_trait>( &'life0 self, request: CallRequest, number: Option<BlockNumber> ) -> Pin<Box<dyn Future<Output = RpcResult<U256>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Estimate gas needed for execution of given contract.

fn gas_price(&self) -> RpcResult<U256>

Returns current gas_price.

fn fee_history<'life0, 'async_trait>( &'life0 self, block_count: U256, newest_block: BlockNumber, reward_percentiles: Option<Vec<f64>> ) -> Pin<Box<dyn Future<Output = RpcResult<FeeHistory>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Introduced in EIP-1159 for getting information on the appropriate priority fee to use.

fn max_priority_fee_per_gas(&self) -> RpcResult<U256>

Introduced in EIP-1159, a Geth-specific and simplified priority fee oracle. Leverages the already existing fee history cache.

fn is_mining(&self) -> RpcResult<bool>

Returns true if client is actively mining new blocks.

fn hashrate(&self) -> RpcResult<U256>

Returns the number of hashes per second that the node is mining with.

fn work(&self) -> RpcResult<Work>

Returns the hash of the current block, the seedHash, and the boundary condition to be met.

fn submit_hashrate(&self, hashrate: U256, id: H256) -> RpcResult<bool>

Used for submitting mining hashrate.

fn submit_work( &self, nonce: H64, pow_hash: H256, mix_digest: H256 ) -> RpcResult<bool>

Used for submitting a proof-of-work solution.

fn send_transaction<'life0, 'async_trait>( &'life0 self, request: TransactionRequest ) -> Pin<Box<dyn Future<Output = RpcResult<H256>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Sends transaction; will block waiting for signer to return the transaction hash.

fn send_raw_transaction<'life0, 'async_trait>( &'life0 self, bytes: Bytes ) -> Pin<Box<dyn Future<Output = RpcResult<H256>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Sends signed transaction, returning its hash.

Provided Methods

fn into_rpc(self) -> RpcModule<Self>

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.