Module fc_rpc_core::types

RPC types

Modules

pubsub
Pub-Sub types.

Structs

AccountInfo
Account information.

Block
Block representation

Bytes
Wrapper structure around vector of bytes.

CallRequest
Call request

ChainStatus
Chain status.

EthAccount
Account information.

EthProtocolInfo
Peer Ethereum protocol information

ExtAccountInfo
Extended account information (used by parity_allAccountInfo).

Filter
Filter

FilterPoolItem

FilteredParams
Helper for Filter matching. Supports conditional indexed parameters and wildcards.

Header
Block header representation.

Index
Represents usize.

Log
Log

PeerInfo
Peer connection information

PeerNetworkInfo
Peer network information

PeerProtocolsInfo
Peer protocols information

Peers
Peers info

PendingTransaction

PipProtocolInfo
Peer PIP protocol information

Receipt
Receipt

RecoveredAccount
account derived from a signature as well as information that tells if it is valid for the current chain

Rich
Value representation with additional info

RichRawTransaction
Geth-compatible output for eth_signTransaction method

StorageProof
Data structure with proof for one single storage-entry

SyncInfo
Sync info

Transaction
Transaction

TransactionRequest
Transaction request coming from RPC

TransactionStats
Propagation statistics for pending transaction.

Work
The result of an eth_getWork call: it differs based on an option whether to send the block number.

Enums

BlockNumber
Represents rpc api block number param.

BlockTransactions
Block Transactions

FilterChanges
Results of the filter_changes RPC.

FilterType

LocalTransactionStatus
Local Transaction Status

SyncStatus
Sync status

VariadicValue
Variadic value

Type Definitions

FilterAddress
Filter Address

FilterPool
On-memory stored filters created through the eth_newFilter RPC.

PendingTransactions

RichBlock
Block representation with additional info.

RichHeader
Header representation with additional info.

Topic
Topic, supports A | null | [A,B,C] | [A,[B,C]] | [null,[B,C]] | [null,[null,C]]